template for arc toolbox
